rename loss_pt to loss
