+++ sockeye/train.py
@@ -37,7 +37,7 @@ from . import constants as C
 from . import data_io
 from . import encoder
 from . import layers
-from . import loss_pt
+from . import loss
 from . import lr_scheduler
 from . import model_pt
 from . import optimizers
